requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
torch
transformers
sentence-transformers
pydantic
python-multipart

